Change catalog endpoint to `/providers/:providerConfigKey/templates`
